[generator] Avoid CS0219 when using [Async] by removing unused var
Currently we get a CS2019 warning whenever we use AsyncAttribute
because we ignore the `result` variable on methods whose return
type is != `void` which is the expected behaviour.

This is specially more annoying on third party bindings,
the generated code did not disable this warning so you get
an "unfixable" warning on the generated code and the user
has no way to fix this on their own.

We now remove the unused `result` varibale to avoid the warning.
Nobody likes warnings that are not their fault.

build diff:
https://gist.github.com/dalexsoto/313f913c394d8eb46f95200524eae2e2