[monotouch-test] Fix SystemSoundTest.FromFile's audio file to not take 23s. Fixes #54236.
In a6b9c28975edcb64da207f48368d051e5c3649a9 I fixed SystemSoundTest.FromFile
to not crash randomly, and at the same time I added an assert to ensure that
the playback completion handler is called properly, with a 10s timeout.

Which won't work, because the audio file is 23s long.

There's no need for a 23s audio file, so the fix is simple: cut the audio file
to 0.3s instead.

https://bugzilla.xamarin.com/show_bug.cgi?id=54236