[tests] Don't create test packages by default. (#1941)
* [tests] Don't create test packages by default.

Don't create test packages by default, instead add a new target to create test
packages. This new target is called on wrench, which means the packages will
still be created when needed, but they won't be built locally in every build
(and if a packaged test fails to build, it won't fail the entire build).

* [tests] Use a project reference instead of assembly reference for GuiUnit.exe

Use a project reference instead of assembly reference for GuiUnit.exe, so that
the GuiUnit reference is automatically built if necessary.

This also makes it required to build a sln for Classic (since mdtool can't
find referenced projects from a csproj).