[tests] Remove OptimizeGeneratedCodeTest.IntPtrSizeOptimization. Fixes #54408
This optimization is now limited to Xamairn.iOS.dll. 3rd parties .dll
size savings were not large enough to justify two different (32/64)
assemblies in the application bundle.

It's likely this change that resulted in incorrect fixes like:

-                       if (classic_or_sim || single_arch) {
+                       if (IsMainExecutableDual ())

which reversed the condition and later
https://github.com/xamarin/xamarin-macios/pull/1872/commits/580f502777b62fc31f79f13e6e611bce5947503b
which disabled DEBUG - but we don't have IL on release so it became
never executed.

reference:
https://bugzilla.xamarin.com/show_bug.cgi?id=54408