[ObjCRuntime] Only use a 'BundleFinder' class when we don't have another internal type we can use. Fixes #56255.
Inside our platform assemblies, we can use the NSObject.NSObject_Disposer
class to find the bundle (this class is not binding an existing Objective-C
class, which would make it unusable for this purpose, and it's also needed
always, which means we're not adding unnecessary bloat to linked platform
assemblies).

However, for System.Net.Http.dll, I couldn't find any publicly visible type
that would fit those criteria, so we're still using a custom 'BundleFinder'
class.

https://bugzilla.xamarin.com/show_bug.cgi?id=56255