[OpenTK] Rework makefiles to avoid zero-length pdbs. (#2136)
For some reason I don't quite understand, I ended up with zero-length pdbs for
OpenTK-1.0.dll locally. Looking at the makefile this might be happening in
custom rules that touch the pdbs (which would cause zero-length files if the
pdb didn't exist before that rule executed), so rewrite the OpenTK targets to
use pattern rules with multiple targets, which tells Make that the same recipe
builds all targets, instead of relying on custom rules that does unrelated
things (i.e. `touch foo.pdb`) to provide a rule for the pdbs.