Keeps references to the delegate passed to Marshal.Marshal.GetFunctionPointerForDelegate
While this issue would have been rare due to the typical scenarios where Blocks are used
and the rare use of CFRunLoop, there is a chance that the delegates that are created
implicitly when passing a method name to GetFunctionPointerForDelegate would be released
and that the native code trying to call back into managed code would have executed
invalid code.

This incorrect use of the GetFunctionPointerForDelegate was discovered
by the TensorFlowSharp bindings on Windows which used a similar idiom.

In the case of CFRunLoop, we now keep the delegates that are used as
static variables in the CFRunLoop class.  On the case of Block,
instead of tracking a single delegate, we now track two delegates in
the GCHandle allcoated object.