[AVFoundation] Adds missing Xcode9 API, fixes AVCaptureSynchronizedDataCollection, improves existing API, and fixes bug 59571
https://bugzilla.xamarin.com/show_bug.cgi?id=59571

Changes in AVFoundation
=======================

AVCaptureSynchronizedDataCollection:

* Adds `NSFastEnumerator` and `IEnumerable<AVCaptureOutput>` support.
* Obsoletes `From` method because the name does not makes sense in the current context.
* Obsoletes `ObjectForKeyedSubscript` in favor of a C# indexer.
* Adds `GetSynchronizedData` to replace obsoleted `From` method`.

AVCaptureSynchronizedData:

* Adds `[Abstract]` in XAMCORE_4_0 because it is an abstract superclass.

AVCapturePhotoOutput:

* Adds missing members `DepthDataDeliverySupported` and `DepthDataDeliveryEnabled`
  from `AVCapturePhotoOutputDepthDataDeliverySupport` category.

AVCaptureDevice:

* Adds strong typed API to `GetAuthorizationStatus` and `RequestAccessForMediaType`
  using `AVAuthorizationMediaType` enum holding the only possible values.
* Adds missing members `ActiveDepthDataFormat`, `MinAvailableVideoZoomFactor` and
  `MaxAvailableVideoZoomFactor` from `AVCaptureDeviceDepthSupport` category.

Changes in Foundation
=====================

NSFastEnumerator:

When implementing `NSFastEnumerator` suppor in `AVCaptureSynchronizedDataCollection`
I ran into a NRE whenever we tried to set the `mutationValue` from the returned state
struct, this is because `Marshal.ReadIntPtr` failed to read `state.mutationsPtr` due to
how `AVCaptureSynchronizedDataCollection` implements `NSFastEnumerator` protocol.

The `AVCaptureSynchronizedDataCollection` implementation stores the value of the number
of items it is holding instead of a pointer to it, this was the reason why `ReadIntPtr` throws.

Now whenever `Marshal.ReadIntPtr` throws we directly use the value of `state.mutationsPtr` as a fallback.

Test
====

A test exercising the new API lives here:

https://github.com/dalexsoto/XamarinTests/blob/4715069b2bd777009c94c32fd9eed56b81103264/AVCaptureDataOutputSynchronizerTest/AVCaptureDataOutputSynchronizerTest/ViewController.cs#L174-L189

Unfortunately the API needs an iPhone 7+, 8+ or X in order to run.